backend/chat_backend/requirements.txt
alabaster==0.7.13
alembic==1.11.2
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==3.0.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.13.1
backcall==0.2.0
black==23.10.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
config==0.5.1
constantly==15.1.0
coverage==7.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.5
cssbeautifier==1.14.11
daphne==4.0.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
Django==4.1.4
django-allauth==0.58.2
django-cors-headers==4.3.0
django-coverage-plugin==3.1.0
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-model-utils==4.3.1
django-redis==5.4.0
django-stubs==4.2.6
django-stubs-ext==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djangorestframework-stubs==3.14.4
djlint==1.34.0
docutils==0.20.1
drf-spectacular==0.26.5
EditorConfig==0.12.3
executing==1.2.0
factory-boy==3.3.0
Faker==15.3.4
filelock==3.12.2
flake8==6.1.0
flake8-isort==6.1.1
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
hiredis==2.2.3
html-tag-names==0.1.2
html-void-elements==0.1.0
hyperlink==21.0.0
identify==2.5.31
idna==3.4
imagesize==1.4.1
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.17.2
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsbeautifier==1.14.11
json5==0.9.14
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
livereload==2.6.3
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
msgpack==1.0.4
mypy==1.6.1
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
mysqlclient==2.2.0
nodeenv==1.8.0
numpy==1.25.1
oauthlib==3.2.2
packaging==23.2
pandas==2.0.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pipenv==2023.10.24
platformdirs==3.9.1
pluggy==1.3.0
pre-commit==3.5.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
Pygments==2.13.0
PyJWT==2.8.0
pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyOpenSSL==23.3.0
pytest==7.4.3
pytest-django==4.6.0
pytest-sugar==0.9.7
python-dateutil==2.8.2
python-decouple==3.7
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0.1
redis==4.4.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.12.0
service-identity==23.1.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.19
sqlparse==0.4.3
stack-data==0.6.2
termcolor==2.3.0
text-unidecode==1.3
tomlkit==0.12.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.6.0
tuna==0.5.11
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==22.2.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
typing_extensions==4.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.24.6
watchdog==3.0.0
wcwidth==0.2.5
Werkzeug==3.0.1
whitenoise==6.6.0
zope.interface==5.5.2
